+++ src/index.ts
@@ -2,6 +2,7 @@ import { Plugin } from 'vite';
 import { Options } from './interfaces/options';
 import { setOptions } from './options';
 import { buildDocumentation } from './searchBuilder';
+import fs from 'fs';
 
 /**
  * Initialize the plugin, and pass additional configuration options.
@@ -35,6 +36,13 @@ export function SimpleSearch(userOptions: Options): Plugin {
         },
         async load(this, id) {
             if (id === resolvedVirtualModuleId) {
+                const filePathway = options.docsPath ? options.docsPath : config.root;
+                if (!fs.existsSync(filePathway)) {
+                    console.warn(`File Pathway: ${filePathway} does not exist. Search could not be built.`)
+                    console.warn(`Try 'process.cwd() + your folder'`)
+                    throw new Error(`Docs pathway could not be found.`);
+                }
+
                 const fileData = await buildDocumentation(config.root);
                 const javaScript: string =
                     `const regexForContentStripping = ${options.regexForContentStripping}` +